+++ tests/integration/dbmultiplerows/Database.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Wikimedia\Rdbms;
+
+class Database {
+	function insert( $table, $vars, $conds = '', $x = 'x', $o = [], $j = [] ) {
+		return (object)[ $_GET['key'] => $_GET['val'] ];
+	}
+}